requirements.txt
fastapi
uvicorn
httpx
python-dotenv
pymongo
python-jose
passlib[bcrypt]
python-multipart
pydantic[email]